Apply glc_nzoc to MPAS-Ocean namelist build system
This commit makes it so MPAS-Ocean gets the value of glc_nzoc from the
value the E3SM bld system assigns it (which comes from the compset
definition).  This commit modifies the files in the MPAS-Ocean namelist
build system, but does not yet alter the way MPAS-Ocean interacts with
coupler at runtime.