Add facemelt and optional ISMF flux to Fogg_rofl
These liquid runoff fluxes are currently added to the river runoff flux
in the coupler before being passed to the ocean.  If we want to change
the horizontal or vertical distribution of how these are handled in the
ocean, we'll need to separate them in the future.  Also needed is time
averaging of fluxes, which has not yet been implemented in MALI.